# specs/001-network-simulator-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: NetLab API
  version: 1.0.0
  description: Shared single-host network-simulator control API.
servers:
  - url: /api/v1
security: []
paths:
...
  /labs/{labId}/nodes:
    parameters: [{$ref: '#/components/parameters/LabId'}]
    post:
      operationId: createNode
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateNode'}}}}
      responses:
        '202': {$ref: '#/components/responses/AcceptedTask'}
...
  /nodes/{nodeId}:
    parameters: [{$ref: '#/components/parameters/NodeId'}]
    get:
      operationId: getNode
      responses:
        '200': {description: Node, content: {application/json: {schema: {$ref: '#/components/schemas/Node'}}}}
    delete:
      operationId: deleteNode
      parameters: [{$ref: '#/components/parameters/IfMatch'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses:
        '202': {$ref: '#/components/responses/AcceptedTask'}
...
  /nodes/{nodeId}/interfaces:
    parameters: [{$ref: '#/components/parameters/NodeId'}]
    post:
      operationId: addInterface
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [driver], properties: {driver: {type: string}}}}}}
      responses:
        '202': {$ref: '#/components/responses/AcceptedTask'}
  /interfaces/{interfaceId}:
    delete:
      operationId: removeInterface
      parameters: [{name: interfaceId, in: path, required: true, schema: {type: string, format: uuid}}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses:
        '202': {$ref: '#/components/responses/AcceptedTask'}
...
  /nodes/{nodeId}/guest-exec:
    parameters: [{$ref: '#/components/parameters/NodeId'}]
    post:
      operationId: executeGuestCommand
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [argv], properties: {argv: {type: array, minItems: 1, items: {type: string}}, timeout_seconds: {type: integer, minimum: 1, maximum: 3600}, output_limit_bytes: {type: integer, minimum: 1024, maximum: 10485760}}}}}}
      responses:
        '202': {$ref: '#/components/responses/AcceptedTask'}
...
  /nodes/{nodeId}/resources:
    get:
      operationId: getNodeResources
      parameters: [{$ref: '#/components/parameters/NodeId'}]
      responses:
        '200': {description: Configured limits and observed cgroup metrics, content: {application/json: {schema: {type: object}}}}
    put:
      operationId: updateNodeResources
      parameters: [{$ref: '#/components/parameters/NodeId'}, {$ref: '#/components/parameters/IfMatch'}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [cpu_count, cpu_quota_micros, memory_mib]}}}}
      responses:
        '200': {description: Updated node, content: {application/json: {schema: {$ref: '#/components/schemas/Node'}}}}
  /nodes/{nodeId}/settings:
    put:
      operationId: updateNodeSettings
      description: Update the name, restart-sensitive resources, interface drivers, and cloud-init network settings of a stopped node.
      parameters: [{$ref: '#/components/parameters/NodeId'}, {$ref: '#/components/parameters/IfMatch'}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, cpu_count, cpu_quota_micros, memory_mib, interface_limit, process_limit]
              properties:
                name: {type: string, minLength: 1, maxLength: 128}
                cpu_count: {type: integer, minimum: 1, maximum: 256}
                cpu_quota_micros: {type: integer, minimum: 0, description: Zero disables CPU-time limiting; 100000 equals one host CPU core per period.}
                memory_mib: {type: integer, minimum: 64}
                interface_limit: {type: integer, minimum: 1}
                process_limit: {type: integer, minimum: 1, maximum: 1048576}
                network_interfaces:
                  type: array
                  items:
                    type: object
                    required: [id, name, driver, modes, addresses]
                    properties:
                      id: {type: string, format: uuid}
                      name: {type: string}
                      driver: {enum: [virtio-net-pci, e1000, e1000e, vmxnet3]}
                      modes: {type: array, items: {enum: [dhcpv4, dhcpv6, slaac, static]}}
                      addresses: {type: array, items: {type: string}}
      responses:
        '200': {description: Updated stopped node, content: {application/json: {schema: {$ref: '#/components/schemas/Node'}}}}
        '409': {description: Node is not stopped, the revision changed, or the name conflicts}
  /nodes/{nodeId}/bootstrap-credentials:
    get:
      operationId: getNodeBootstrapCredentials
      description: Recover the initial username and plaintext password from a managed Ubuntu cloud-init seed ISO.
      parameters: [{$ref: '#/components/parameters/NodeId'}]
      responses:
        '200':
          description: Recoverable initial cloud-init credentials
          content:
            application/json:
              schema:
                type: object
                required: [username, password, source]
                properties:
                  username: {type: string}
                  password: {type: string}
                  source: {type: string, enum: [cloud-init-seed]}
        '404': {description: The node has no recoverable Ubuntu cloud-init credentials}
...
components:
  parameters:
    LabId: {name: labId, in: path, required: true, schema: {type: string, format: uuid}}
    NodeId: {name: nodeId, in: path, required: true, schema: {type: string, format: uuid}}
...
    IfMatch: {name: If-Match, in: header, required: true, schema: {type: string}, description: Quoted resource revision}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: false, schema: {type: string, minLength: 8, maxLength: 128}}
  responses:
    AcceptedTask: {description: Accepted, content: {application/json: {schema: {$ref: '#/components/schemas/TaskEnvelope'}}}}
...
  schemas:
...
    Node:
      type: object
      required: [id, laboratory_id, name, kind, revision, desired_state, observed_state]
      properties: {id: {type: string, format: uuid}, laboratory_id: {type: string, format: uuid}, name: {type: string}, kind: {enum: [qemu, docker, pc, switch_l2, switch_l3]}, template_version_id: {type: string, format: uuid}, revision: {type: integer}, desired_state: {enum: [stopped, running, deleted]}, observed_state: {enum: [unknown, provisioning, starting, running, stopping, stopped, failed, deleting]}, cpu_count: {type: integer}, cpu_quota_micros: {type: integer}, memory_mib: {type: integer}, storage_gib: {type: integer}, interface_limit: {type: integer}, process_limit: {type: integer}, config: {type: object}, last_error: {$ref: '#/components/schemas/Problem'}}
    CreateNode:
      type: object
      required: [name]
      properties: {name: {type: string}, kind: {enum: [qemu, docker, pc, switch_l2, switch_l3]}, template_version_id: {type: string, format: uuid}, image_version_id: {type: string, format: uuid}, cpu_count: {type: integer}, cpu_quota_micros: {type: integer}, memory_mib: {type: integer}, storage_gib: {type: integer}, interface_limit: {type: integer}, process_limit: {type: integer}, interface_count: {type: integer}, config: {type: object}, bootstrap: {type: object, properties: {user_data: {type: string}, meta_data: {type: string}, network_config: {type: string}}}}
    NodeInterface:
      type: object
      required: [id, node_id, slot, name, driver, mac_address, revision]
      properties: {id: {type: string, format: uuid}, node_id: {type: string, format: uuid}, slot: {type: integer}, name: {type: string}, driver: {type: string}, mac_address: {type: string}, desired_link_id: {type: string, format: uuid}, operational_state: {type: string}, revision: {type: integer}}
    Link:
      type: object
      required: [id, laboratory_id, endpoint_a_id, endpoint_b_id, revision, desired_state, observed_state]
      properties: {id: {type: string, format: uuid}, laboratory_id: {type: string, format: uuid}, endpoint_a_id: {type: string, format: uuid}, endpoint_b_id: {type: string, format: uuid}, revision: {type: integer}, desired_state: {type: string}, observed_state: {type: string}}
...
    PortMapping:
      type: object
      required: [id, node_id, protocol, host_address, host_port, guest_address, guest_port, revision, observed_state, created_at]
      properties: {id: {type: string, format: uuid}, node_id: {type: string, format: uuid}, protocol: {enum: [tcp, udp]}, host_address: {type: string}, host_port: {type: integer, minimum: 1, maximum: 65535}, guest_address: {type: string}, guest_port: {type: integer, minimum: 1, maximum: 65535}, revision: {type: integer, minimum: 1}, observed_state: {enum: [pending, active, failed, deleting]}, last_error: {$ref: '#/components/schemas/Problem'}, created_at: {type: string, format: date-time}}
...
    OperationTask:
      type: object
      required: [id, kind, resource_type, resource_id, state, progress_current, progress_total, created_at]
      properties: {id: {type: string, format: uuid}, kind: {type: string}, resource_type: {type: string}, resource_id: {type: string}, requested_revision: {type: integer}, state: {enum: [queued, running, cancelling, succeeded, failed, cancelled]}, progress_current: {type: integer}, progress_total: {type: integer}, result: {type: object}, error: {$ref: '#/components/schemas/Problem'}, created_at: {type: string, format: date-time}, started_at: {type: string, format: date-time}, finished_at: {type: string, format: date-time}}
...
    TaskEnvelope:
      type: object
      required: [task]
      properties: {task: {$ref: '#/components/schemas/OperationTask'}, node: {$ref: '#/components/schemas/Node'}, link: {$ref: '#/components/schemas/Link'}, interface: {$ref: '#/components/schemas/NodeInterface'}, port_mapping: {$ref: '#/components/schemas/PortMapping'}, artifact: {$ref: '#/components/schemas/Artifact'}}
    Problem:
      type: object
      properties: {code: {type: string}, message: {type: string}, retryable: {type: boolean}, resource_type: {type: string}, resource_id: {type: string}, task_id: {type: string, format: uuid}, phase: {type: string}, cleanup: {type: string}, operator_hint: {type: string}, retry_after_seconds: {type: integer, minimum: 1}, details: {type: object}}